requirements.txt
numpy>=1.20.0
sounddevice>=0.4.5
openai-whisper>=20231117
PyQt5>=5.15.0
pyttsx3>=2.90
PyAudio>=0.2.11
torch>=2.0.0
torchaudio>=2.0.0